+++ src/shared/ui/form-components/controlled-select/index.ts
@@ -1,0 +1,1 @@
+export * from 'src/shared/ui/form-components/controlled-select/controlled-select'
