--- src/shared/ui/from-components/controlled-select/index.ts
@@ -1,1 +1,0 @@
-export * from './controlled-select'
